Publish Choice.successChance as double
Previously, it was converted to 1..100 integer, but that was just annoying.

+++ edgehead/lib/egamebook/elements/choice_element.dart
@@ -18,8 +18,8 @@ abstract class Choice extends ElementBase
 
   String get command;
 
-  /// The chance of success, from `0` to `100`.
-  int get successChance;
+  /// The chance of success, from `0` to `1`.
+  double get successChance;
 
   BuiltList<String> get commandPath;
